requirements.txt
streamlit
torch
whisper
transformers
easyocr
pillow
imageio-ffmpeg
google-generativeai
opencv-python
